requirements.txt
Django~=4.2.6
Pillow~=10.1.0
django-autoslug~=1.9.9
django-debug-toolbar~=4.2.0
social-auth-app-django==5.4.0
channels~=4.0.0
django-extensions~=3.2.3
daphne~=4.0.0
asgiref~=3.7.2
asgi_redis~=1.4.3
channels-redis~=4.1.0
djangorestframework~=3.14.0
django-imagekit~=5.0.0
djoser~=2.2.2